+++ srcs/parser/substitution/tilda.c
@@ -1,19 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tilda.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rbednar <rbednar@student.21-school.ru>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/07/25 16:02:18 by rbednar           #+#    #+#             */
+/*   Updated: 2020/08/01 15:23:10 by rbednar          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "shell42.h"
 #include "parser.h"
 
 int		ft_find_tilda(t_ltree *sub, int flag)
 {
-	size_t	i;
+	int	i;
 
 	i = -1;
-	while (++i < sub->l_tline.len)
+	while ((size_t)++i < sub->l_tline.len)
 	{
 		if ((i == 0 || (i > 0 && (sub->l_tline.line[i - 1] == SPACE ||
 			(flag == ASSIGN && (sub->l_cmd[i - 1] == ':' ||
 			sub->l_tline.line[i - 1] == EQUAL))))) &&
 			sub->l_tline.line[i] == TILDA)
 		{
-			if (i + 1 == sub->l_tline.len || sub->l_tline.line[i + 1] == ENTER ||
+			if ((size_t)(i + 1) == sub->l_tline.len ||
+				sub->l_tline.line[i + 1] == ENTER ||
 				sub->l_tline.line[i + 1] == SPACE ||
 				sub->l_cmd[i + 1] == '/' ||
 				(flag == ASSIGN && sub->l_cmd[i + 1] == ':'))
@@ -29,9 +42,9 @@ int		ft_find_tilda(t_ltree *sub, int flag)
 ** It works when no symbol after '~' or it's '/'
 */
 
-int 	ft_get_home(t_ltree *sub, size_t *i)
+int		ft_get_home(t_ltree *sub, int *i)
 {
-	char    *tmp;
+	char	*tmp;
 
 	if ((tmp = ft_strdup(find_env_value("HOME"))) != NULL)
 	{
@@ -51,13 +64,13 @@ int 	ft_get_home(t_ltree *sub, size_t *i)
 }
 
 /*
-** It works when after symbol '~' is some WORD 
+** It works when after symbol '~' is some WORD
 ** until '/' or ':' (only if flag == ASSIGN)
 */
 
-int	ft_getdir_by_name(t_ltree *sub, size_t *i, int flag)
+int		ft_getdir_by_name(t_ltree *sub, int *i, int flag)
 {
-	size_t	j;
+	int		j;
 	char	*user;
 
 	j = 1;
@@ -75,12 +88,12 @@ int	ft_getdir_by_name(t_ltree *sub, size_t *i, int flag)
 ** Function open file /etc/passwd and try to find user in it by USER
 */
 
-int		ft_find_dir_info(t_ltree *sub, char *user, size_t *i)
+int		ft_find_dir_info(t_ltree *sub, char *user, int *i)
 {
 	int		fd;
 	char	*line;
 	char	**info;
-	
+
 	line = NULL;
 	if ((fd = open("/etc/passwd", O_RDONLY)) != -1)
 		while (ft_gnl(fd, &line) > 0)
@@ -97,7 +110,7 @@ int		ft_find_dir_info(t_ltree *sub, char *user, size_t *i)
 			}
 			ft_arrdel(info);
 		}
-	close (fd);
+	close(fd);
 	return (0);
 }
 
@@ -105,12 +118,12 @@ int		ft_find_dir_info(t_ltree *sub, char *user, size_t *i)
 ** Function open file /etc/passwd and try to find user in it by UID
 */
 
-int		ft_find_dir_by_uid(t_ltree *sub, char *uid, size_t *i)
+int		ft_find_dir_by_uid(t_ltree *sub, char *uid, int *i)
 {
 	int		fd;
 	char	*line;
 	char	**info;
-	
+
 	line = NULL;
 	if ((fd = open("/etc/passwd", O_RDONLY)) != -1)
 		while (ft_gnl(fd, &line) > 0)
@@ -127,6 +140,6 @@ int		ft_find_dir_by_uid(t_ltree *sub, char *uid, size_t *i)
 			}
 			ft_arrdel(info);
 		}
-	close (fd);
+	close(fd);
 	return (0);
 }
